Implement Wave Management System
Add `Wave` base class with an abstract `CheckCondition` method for derived wave classes to implement.
Implement `EnemyWave` class that triggers the next wave when all its child enemies are disabled.
Modify `EnableNextWave` method to start at wave 0 and add safety checks for array bounds.
Add a warning message when there are no more waves to enable.
Created WaveManager_Food prefab for copying into Food scene, as well as to work as a template for the system.